+++ lite3dpp/src/json/JSON.cpp
@@ -46,31 +46,31 @@ JSON::JSON()
  */
 JSONValue *JSON::Parse(const char *data)
 {
-	size_t length = strlen(data) + 1;
-	wchar_t *w_data = (wchar_t*)malloc(length * sizeof(wchar_t));
-	
-	#if defined(WIN32) && !defined(__GNUC__)
-		size_t ret_value = 0;
-		if (mbstowcs_s(&ret_value, w_data, length, data, length) != 0)
-		{
-			free(w_data);
-			return NULL;
-		}
-	#elif defined(ANDROID)
-		// mbstowcs seems to misbehave on android
-		for(size_t i = 0; i<length; i++)
-			w_data[i] = (wchar_t)data[i];
-	#else
-		if (mbstowcs(w_data, data, length) == (size_t)-1)
-		{
-			free(w_data);
-			return NULL;
-		}
-	#endif
-	
-	JSONValue *value = JSON::Parse(w_data);
-	free(w_data);
-	return value;
+    size_t length = strlen(data) + 1;
+    wchar_t *w_data = (wchar_t*)malloc(length * sizeof(wchar_t));
+    
+    #if defined(WIN32) && !defined(__GNUC__)
+        size_t ret_value = 0;
+        if (mbstowcs_s(&ret_value, w_data, length, data, length) != 0)
+        {
+            free(w_data);
+            return NULL;
+        }
+    #elif defined(ANDROID)
+        // mbstowcs seems to misbehave on android
+        for(size_t i = 0; i<length; i++)
+            w_data[i] = (wchar_t)data[i];
+    #else
+        if (mbstowcs(w_data, data, length) == (size_t)-1)
+        {
+            free(w_data);
+            return NULL;
+        }
+    #endif
+    
+    JSONValue *value = JSON::Parse(w_data);
+    free(w_data);
+    return value;
 }
 
 /**
@@ -84,24 +84,24 @@ JSONValue *JSON::Parse(const char *data)
  */
 JSONValue *JSON::Parse(const wchar_t *data)
 {
-	// Skip any preceding whitespace, end of data = no JSON = fail
-	if (!SkipWhitespaceAndComments(&data))
-		return NULL;
+    // Skip any preceding whitespace, end of data = no JSON = fail
+    if (!SkipWhitespaceAndComments(&data))
+        return NULL;
 
-	// We need the start of a value here now...
-	JSONValue *value = JSONValue::Parse(&data);
-	if (value == NULL)
-		return NULL;
-	
-	// Can be white space now and should be at the end of the string then...
-	if (SkipWhitespaceAndComments(&data))
-	{
-		delete value;
-		return NULL;
-	}
-	
-	// We're now at the end of the string
-	return value;
+    // We need the start of a value here now...
+    JSONValue *value = JSONValue::Parse(&data);
+    if (value == NULL)
+        return NULL;
+    
+    // Can be white space now and should be at the end of the string then...
+    if (SkipWhitespaceAndComments(&data))
+    {
+        delete value;
+        return NULL;
+    }
+    
+    // We're now at the end of the string
+    return value;
 }
 
 /**
@@ -115,10 +115,10 @@ JSONValue *JSON::Parse(const wchar_t *data)
  */
 lite3dpp::WString JSON::Stringify(const JSONValue *value)
 {
-	if (value != NULL)
-		return value->Stringify();
-	else
-		return L"";
+    if (value != NULL)
+        return value->Stringify();
+    else
+        return L"";
 }
 
 /**
@@ -132,10 +132,10 @@ lite3dpp::WString JSON::Stringify(const JSONValue *value)
  */
 bool JSON::SkipWhitespace(const wchar_t **data)
 {
-	while (**data != 0 && (**data == L' ' || **data == L'\t' || **data == L'\r' || **data == L'\n'))
-		(*data)++;
-	
-	return **data != 0;
+    while (**data != 0 && (**data == L' ' || **data == L'\t' || **data == L'\r' || **data == L'\n'))
+        (*data)++;
+    
+    return **data != 0;
 }
 
 bool JSON::SkipWhitespaceAndComments(const wchar_t **data)
@@ -167,92 +167,92 @@ bool JSON::SkipWhitespaceAndComments(const wchar_t **data)
  */
 bool JSON::ExtractString(const wchar_t **data, lite3dpp::WString &str)
 {
-	str = L"";
-	
-	while (**data != 0)
-	{
-		// Save the char so we can change it if need be
-		wchar_t next_char = **data;
-		
-		// Escaping something?
-		if (next_char == L'\\')
-		{
-			// Move over the escape char
-			(*data)++;
-			
-			// Deal with the escaped char
-			switch (**data)
-			{
-				case L'"': next_char = L'"'; break;
-				case L'\\': next_char = L'\\'; break;
-				case L'/': next_char = L'/'; break;
-				case L'b': next_char = L'\b'; break;
-				case L'f': next_char = L'\f'; break;
-				case L'n': next_char = L'\n'; break;
-				case L'r': next_char = L'\r'; break;
-				case L't': next_char = L'\t'; break;
-				case L'u':
-				{
-					// We need 5 chars (4 hex + the 'u') or its not valid
-					if (!simplejson_wcsnlen(*data, 5))
-						return false;
-					
-					// Deal with the chars
-					next_char = 0;
-					for (int i = 0; i < 4; i++)
-					{
-						// Do it first to move off the 'u' and leave us on the
-						// final hex digit as we move on by one later on
-						(*data)++;
-						
-						next_char <<= 4;
-						
-						// Parse the hex digit
-						if (**data >= '0' && **data <= '9')
-							next_char |= (**data - '0');
-						else if (**data >= 'A' && **data <= 'F')
-							next_char |= (10 + (**data - 'A'));
-						else if (**data >= 'a' && **data <= 'f')
-							next_char |= (10 + (**data - 'a'));
-						else
-						{
-							// Invalid hex digit = invalid JSON
-							return false;
-						}
-					}
-					break;
-				}
-				
-				// By the spec, only the above cases are allowed
-				default:
-					return false;
-			}
-		}
-		
-		// End of the string?
-		else if (next_char == L'"')
-		{
-			(*data)++;
-			str.reserve(); // Remove unused capacity
-			return true;
-		}
-		
-		// Disallowed char?
-		else if (next_char < L' ' && next_char != L'\t')
-		{
-			// SPEC Violation: Allow tabs due to real world cases
-			return false;
-		}
-		
-		// Add the next char
-		str += next_char;
-		
-		// Move on
-		(*data)++;
-	}
-	
-	// If we're here, the string ended incorrectly
-	return false;
+    str = L"";
+    
+    while (**data != 0)
+    {
+        // Save the char so we can change it if need be
+        wchar_t next_char = **data;
+        
+        // Escaping something?
+        if (next_char == L'\\')
+        {
+            // Move over the escape char
+            (*data)++;
+            
+            // Deal with the escaped char
+            switch (**data)
+            {
+                case L'"': next_char = L'"'; break;
+                case L'\\': next_char = L'\\'; break;
+                case L'/': next_char = L'/'; break;
+                case L'b': next_char = L'\b'; break;
+                case L'f': next_char = L'\f'; break;
+                case L'n': next_char = L'\n'; break;
+                case L'r': next_char = L'\r'; break;
+                case L't': next_char = L'\t'; break;
+                case L'u':
+                {
+                    // We need 5 chars (4 hex + the 'u') or its not valid
+                    if (!simplejson_wcsnlen(*data, 5))
+                        return false;
+                    
+                    // Deal with the chars
+                    next_char = 0;
+                    for (int i = 0; i < 4; i++)
+                    {
+                        // Do it first to move off the 'u' and leave us on the
+                        // final hex digit as we move on by one later on
+                        (*data)++;
+                        
+                        next_char <<= 4;
+                        
+                        // Parse the hex digit
+                        if (**data >= '0' && **data <= '9')
+                            next_char |= (**data - '0');
+                        else if (**data >= 'A' && **data <= 'F')
+                            next_char |= (10 + (**data - 'A'));
+                        else if (**data >= 'a' && **data <= 'f')
+                            next_char |= (10 + (**data - 'a'));
+                        else
+                        {
+                            // Invalid hex digit = invalid JSON
+                            return false;
+                        }
+                    }
+                    break;
+                }
+                
+                // By the spec, only the above cases are allowed
+                default:
+                    return false;
+            }
+        }
+        
+        // End of the string?
+        else if (next_char == L'"')
+        {
+            (*data)++;
+            str.reserve(); // Remove unused capacity
+            return true;
+        }
+        
+        // Disallowed char?
+        else if (next_char < L' ' && next_char != L'\t')
+        {
+            // SPEC Violation: Allow tabs due to real world cases
+            return false;
+        }
+        
+        // Add the next char
+        str += next_char;
+        
+        // Move on
+        (*data)++;
+    }
+    
+    // If we're here, the string ended incorrectly
+    return false;
 }
 
 /**
@@ -262,15 +262,15 @@ bool JSON::ExtractString(const wchar_t **data, lite3dpp::WString &str)
  *
  * @param wchar_t** data Pointer to a wchar_t* that contains the JSON text
  *
- * @return double Returns the double value of the number found
+ * @return float Returns the float value of the number found
  */
-double JSON::ParseInt(const wchar_t **data)
+float JSON::ParseInt(const wchar_t **data)
 {
-	double integer = 0;
-	while (**data != 0 && **data >= '0' && **data <= '9')
-		integer = integer * 10 + (*(*data)++ - '0');
-	
-	return integer;
+    float integer = 0;
+    while (**data != 0 && **data >= '0' && **data <= '9')
+        integer = integer * 10 + (*(*data)++ - '0');
+    
+    return integer;
 }
 
 /**
@@ -280,19 +280,19 @@ double JSON::ParseInt(const wchar_t **data)
  *
  * @param wchar_t** data Pointer to a wchar_t* that contains the JSON text
  *
- * @return double Returns the double value of the decimal found
+ * @return float Returns the float value of the decimal found
  */
-double JSON::ParseDecimal(const wchar_t **data)
+float JSON::ParseDecimal(const wchar_t **data)
 {
-	double decimal = 0.0;
-  double factor = 0.1;
-	while (**data != 0 && **data >= '0' && **data <= '9')
+    float decimal = 0.0;
+    float factor = 0.1;
+    while (**data != 0 && **data >= '0' && **data <= '9')
   {
     int digit = (*(*data)++ - '0');
-		decimal = decimal + digit * factor;
+        decimal = decimal + digit * factor;
     factor *= 0.1;
   }
-	return decimal;
+    return decimal;
 }
 
 lite3dpp::String JSON::wStringToString(const lite3dpp::WString &str)
